Refactor: Consolidate organization controller and routing logic
